requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
argcomplete==1.10.3
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
Automat==22.10.0
beautifulsoup4==4.8.2
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
comm==0.2.2
compressed_rtf==1.0.6
constantly==23.10.4
cryptography==42.0.5
cssselect==1.2.0
debugpy==1.8.1
decorator==5.1.1
Django==5.0.4
djangorestframework==3.15.1
docx2txt==0.8
ebcdic==1.1.1
executing==2.0.1
extract-msg==0.28.7
filelock==3.13.1
frozenlist==1.4.1
hyperlink==21.0.0
idna==3.6
IMAPClient==2.1.0
incremental==22.10.0
ipykernel==6.29.4
ipython==8.22.2
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.19.1
jmespath==1.0.1
jupyter_client==8.6.1
jupyter_core==5.7.2
lxml==5.1.0
matplotlib-inline==0.1.6
multidict==6.0.5
neo4j==5.19.0
nest-asyncio==1.6.0
numpy==1.26.4
olefile==0.47
packaging==23.2
pandas==2.2.1
parsel==1.8.1
parso==0.8.3
pdfminer.six==20191110
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
Protego==0.3.0
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
PyDispatcher==2.0.7
Pygments==2.17.2
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pyOpenSSL==24.0.0
PyPDF2==3.0.1
pysolr==3.9.0
python-dateutil==2.9.0.post0
python-pptx==0.6.23
pytz==2024.1
pyzmq==25.1.2
queuelib==1.6.2
requests==2.31.0
requests-file==2.0.0
Scrapy==2.11.1
service-identity==24.1.0
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
sqlparse==0.5.0
stack-data==0.6.3
textract==1.6.5
tldextract==5.1.1
tornado==6.4
traitlets==5.14.1
Twisted==24.3.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
w3lib==2.1.2
wcwidth==0.2.13
xlrd==1.2.0
XlsxWriter==3.2.0
yarl==1.9.4
zope.interface==6.2
